+++ rotkehlchen/chain/evm/decoding/balancer/constants.py
@@ -10,6 +10,19 @@ CPT_BALANCER_V1: Final = 'balancer-v1'
 CPT_BALANCER_V2: Final = 'balancer-v2'
 CPT_BALANCER_V3: Final = 'balancer-v3'
 
+CPT_BEETS_V2: Final = 'beets-v2'
+CPT_BEETS_V3: Final = 'beets-v3'
+CPT_BEETS_SWAP_V3: Final = 'beets-swap-v3'
+BEETS_LABEL: Final = 'Beets'
+
+BalancerCounterparty = Literal[
+    'balancer-v1',
+    'balancer-v2',
+    'balancer-v3',
+    'beets-v2',
+    'beets-v3',
+]
+
 BALANCER_API_CHUNK_SIZE: Final = 100
 BALANCER_API_URL: Final = 'https://api-v3.balancer.fi'
 CHAIN_ID_TO_BALANCER_API_MAPPINGS: Final = {
@@ -21,16 +34,21 @@ CHAIN_ID_TO_BALANCER_API_MAPPINGS: Final = {
     ChainID.OPTIMISM: 'OPTIMISM',
     ChainID.POLYGON_POS: 'POLYGON',
     ChainID.ARBITRUM_ONE: 'ARBITRUM',
+    ChainID.SONIC: 'SONIC',
 }
-BALANCER_CACHE_TYPE_MAPPING: Final = {
+BALANCER_CACHE_TYPE_MAPPING: Final[dict[BalancerCounterparty, Literal[CacheType.BALANCER_V1_POOLS, CacheType.BALANCER_V2_POOLS, CacheType.BALANCER_V3_POOLS]]] = {  # noqa: E501
     CPT_BALANCER_V1: CacheType.BALANCER_V1_POOLS,
     CPT_BALANCER_V2: CacheType.BALANCER_V2_POOLS,
     CPT_BALANCER_V3: CacheType.BALANCER_V3_POOLS,
+    CPT_BEETS_V2: CacheType.BALANCER_V2_POOLS,
+    CPT_BEETS_V3: CacheType.BALANCER_V3_POOLS,
 }
-BALANCER_VERSION_MAPPING: Final[dict[str, Literal[1, 2, 3]]] = {
+BALANCER_VERSION_MAPPING: Final[dict[BalancerCounterparty, Literal[1, 2, 3]]] = {
     CPT_BALANCER_V1: 1,
     CPT_BALANCER_V2: 2,
     CPT_BALANCER_V3: 3,
+    CPT_BEETS_V2: 2,
+    CPT_BEETS_V3: 3,
 }
 
 GET_POOLS_QUERY: Final = """