+++ rotkehlchen/tests/utils/decoders.py
@@ -6,6 +6,8 @@ from rotkehlchen.chain.evm.decoding.balancer.constants import (
     CPT_BALANCER_V1,
     CPT_BALANCER_V2,
     CPT_BALANCER_V3,
+    CPT_BEETS_V2,
+    CPT_BEETS_V3,
 )
 from rotkehlchen.chain.evm.decoding.curve.constants import CPT_CURVE
 from rotkehlchen.chain.evm.decoding.extrafi.constants import CPT_EXTRAFI
@@ -47,7 +49,7 @@ def patch_decoder_reload_data(load_global_caches: list[str] | None = None) -> Ex
             ({CPT_GEARBOX}, 'rotkehlchen.chain.evm.decoding.gearbox.gearbox_cache', True, False),
             ({CPT_AERODROME, CPT_VELODROME}, 'rotkehlchen.chain.evm.decoding.velodrome.velodrome_cache', True, False),  # noqa: E501
             ({CPT_EXTRAFI}, 'rotkehlchen.chain.evm.decoding.extrafi.cache', True, True),
-            ({CPT_BALANCER_V1, CPT_BALANCER_V2, CPT_BALANCER_V3}, 'rotkehlchen.chain.evm.decoding.balancer.balancer_cache', True, False),  # noqa: E501
+            ({CPT_BALANCER_V1, CPT_BALANCER_V2, CPT_BALANCER_V3, CPT_BEETS_V2, CPT_BEETS_V3}, 'rotkehlchen.chain.evm.decoding.balancer.balancer_cache', True, False),  # noqa: E501
         ):
             if load_global_caches is not None and any(cache in counterparties for cache in load_global_caches):  # noqa: E501
                 continue